feat(#6) medication -> 약 조회, healtcare : 생성 api, 전체,개별조회 api , js 파일 분리,, healthcare,medicalcare서비스단 생성
