+++ app/routes/dailycare/dailycare_api.py
@@ -1,7 +1,7 @@
-from flask import Flask, Blueprint, jsonify, render_template
+from flask import Flask, Blueprint, jsonify, render_template, request
 from app.services.pet_service import PetService
-
-
+from app.services.dailycare.healthcare_service import HealthCareService
+from app.services.dailycare.medicalcare_service import MedicationService
 import logging
 
 dailycare_api_bp = Blueprint('dailycare_api_bp', __name__, url_prefix= '/api/dailycares')
@@ -16,6 +16,7 @@ def load_modal(name):
 
 logging.basicConfig(level=logging.INFO)
 
+# 회원의 전체 펫 조회
 @dailycare_api_bp.route("/get-pet/<user_id>")
 def get_my_pet(user_id):
     pets = PetService.get_pets_by_user(user_id)
@@ -30,6 +31,7 @@ def get_my_pet(user_id):
     
     return jsonify(pets.to_dict()), 200
 
+# 개별 펫 조회
 @dailycare_api_bp.route('/get-pet/<user_id>/<pet_id>')
 def get_my_pet_by_user_id(user_id,pet_id):
     print(f'입력된 petId : {pet_id} userId {user_id}')
@@ -41,3 +43,65 @@ def get_my_pet_by_user_id(user_id,pet_id):
     
     return jsonify(pets.to_dict()), 200
 
+# HealtCare 생성
+@dailycare_api_bp.route('/save/healthcare/<pet_id>', methods = ['POST'])
+def save_healthcare(pet_id):
+    data = request.json
+    record = HealthCareService.create_health_record(
+        pet_id=pet_id,
+        food=data.get('food'),
+        water=data.get('water'),
+        excrement_status=data.get('excrement_status'),
+        weight_kg=data.get('weight_kg'),
+        walk_time_minutes=data.get('walk_time_minutes'),
+    )
+     # 2. Medication 연결 (여러 개 가능)
+    medication_ids = data.get('medication_ids', [])
+    HealthCareService.link_medications(record.care_id, medication_ids)
+    return jsonify(record.to_dict()), 201
+
+#Healthcare조회
+@dailycare_api_bp.route('/healthcare/<pet_id>' )
+def get_health_log(pet_id):
+    log = HealthCareService.get_health_records_by_pet(pet_id)
+    if isinstance(log, list):
+        return jsonify([l.to_dict() for l in log]), 200
+    else:
+        return jsonify(log.to_dict()), 200
+
+
+# HealthCare 단일 조회
+@dailycare_api_bp.route('/healthcare/<int:pet_id>/<int:care_id>', methods=['GET'])
+def get_health_record(pet_id, care_id):
+    record = HealthCareService.get_health_record_by_id(care_id, pet_id)
+    if not record:
+        return jsonify({"error": "Health record not found"}), 404
+    return jsonify(record.to_dict()), 200
+
+# Medication 조회
+@dailycare_api_bp.route('/medications/<int:pet_id>', methods=['GET'])
+def get_medications(pet_id):
+    meds = MedicationService.get_medications_by_pet(pet_id)
+    return jsonify([m.to_dict() for m in meds]), 200
+
+@dailycare_api_bp.route('/get-healthcare/<int:care_id>', methods=['GET'])
+def get_healthcare(care_id):
+    record = HealthCareService.get_health_record_by_id(care_id)
+    if not record:
+        return jsonify({"error": "Record not found"}), 404
+
+    # 연결된 Medication 포함
+    meds = [
+        {
+            "medication_id": link.medication.medication_id,
+            "name": link.medication.medication_name,
+            "dosage": link.medication.dosage,
+            "frequency": link.medication.frequency
+        }
+        for link in record.medication_links
+    ]
+
+    result = record.to_dict()
+    result['medications'] = meds
+
+    return jsonify(result), 200
